requirements.txt
Flask==3.0.3
requests==2.32.3
beautifulsoup4==4.12.3
lxml==5.3.0
requests-cache==1.2.1
python-dotenv==1.0.1